+++ simulators/lean/helper/lean_spec_client_runner.py
@@ -6,64 +6,56 @@ import asyncio
 import gc
 import importlib
 import inspect
-import io
-import ipaddress
 import json
 import logging
 import os
-import shutil
 import sys
-import time
-import types
 from contextlib import suppress
-from pathlib import Path
 from typing import Any, Final
 
-from aiohttp import web
-from Crypto.Hash import keccak
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.asymmetric import ec
-from cryptography.hazmat.primitives.asymmetric.utils import (
-    Prehashed,
-    decode_dss_signature,
-    encode_dss_signature,
+from lean_spec_runtime import (
+    ATTESTATION_SECRET_FIELD,
+    Checkpoint,
+    HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE,
+    PROPOSAL_SECRET_FIELD,
+    ValidatorIndex,
+    api_server_supports_signed_block_getter,
+    build_helper_bootnode_enr,
+    build_node_config,
+    cache_signed_block,
+    configure_event_source_network,
+    decode_secret_key,
+    dial_bootnodes,
+    extract_inner_block,
+    helper_genesis_metadata,
+    identity_keypair_from_private_key_hex,
+    install_fast_ssz_deserialization,
+    install_legacy_signed_block_wire_compatibility,
+    install_low_s_identity_signature_compatibility,
+    install_setup_prover_compatibility,
+    install_snappy_compress_fallback,
+    keep_status_current,
+    legacy_signed_block_compatibility_enabled,
+    load_genesis,
+    load_validator,
+    load_validator_registry_from_manifest,
+    parse_bootnodes,
+    parse_validator_indices,
+    prepare_runtime_assets,
+    refresh_status,
+    setup_logging,
+    start_listener_and_gossipsub,
+    start_metadata_server,
+    subscribe_gossip_topics,
+    uses_latest_leanspec_format,
+    wire_compat_mode,
 )
-import yaml
-
-
-def install_numba_njit_stub() -> None:
-    if os.environ.get("HIVE_LEAN_SPEC_USE_NUMBA", "0") == "1" or "numba" in sys.modules:
-        return
-
-    numba_module = types.ModuleType("numba")
-
-    def njit(*args: object, **_: object) -> object:
-        if len(args) == 1 and callable(args[0]):
-            return args[0]
-
-        def decorate(func: object) -> object:
-            return func
-
-        return decorate
-
-    numba_module.njit = njit
-    sys.modules["numba"] = numba_module
-
-
-install_numba_njit_stub()
-
 
 from lean_spec.subspecs.api import ApiServer, ApiServerConfig
-from lean_spec.subspecs.chain.config import ATTESTATION_COMMITTEE_COUNT
-from lean_spec.subspecs.genesis.config import GenesisConfig
 from lean_spec.subspecs.metrics import registry as metrics
 from lean_spec.subspecs.networking.client import LiveNetworkEventSource
-from lean_spec.subspecs.networking.enr import ENR, keys as enr_keys
-from lean_spec.subspecs.networking.gossipsub import GossipTopic
-from lean_spec.subspecs.networking.reqresp.message import Status
 from lean_spec.subspecs.networking.service.events import GossipBlockEvent
 from lean_spec.subspecs.networking.transport.identity import IdentityKeypair
-from lean_spec.subspecs.networking.transport.identity.keypair import Secp256k1PublicKey
 from lean_spec.subspecs.networking.transport.quic.connection import (
     QuicConnectionManager,
     QuicStream,
@@ -79,56 +71,24 @@ except ImportError:
 from lean_spec.subspecs.ssz.hash import hash_tree_root
 from lean_spec.subspecs.validator import ValidatorRegistry
 from lean_spec.subspecs.validator.registry import ValidatorEntry
-from lean_spec.subspecs.xmss import SecretKey
 from lean_spec.types import Bytes32
 try:
     from lean_spec.types.participation import AggregationBits, ValidatorIndices
 except ImportError:
     from lean_spec.types import AggregationBits, ValidatorIndices
-from lean_spec.types.collections import SSZList, SSZVector, _validate_offsets
-from lean_spec.types.constants import OFFSET_BYTE_LENGTH
-from lean_spec.types.container import Container
-from lean_spec.types.exceptions import SSZSerializationError, SSZValueError
 from lean_spec.types.uint import Uint32, Uint64
-from lean_spec_runtime import (
-    Checkpoint,
-    SubnetId,
-    ValidatorIndex,
-    build_node_config,
-    configure_event_source_network,
-)
 
 DEFAULT_GOSSIP_FORK_DIGEST: Final = "devnet0"
 DEFAULT_LISTEN_PORT: Final = 9001
 DEFAULT_API_PORT: Final = 5052
 DEFAULT_HELPER_METADATA_PORT: Final = 5053
-BOOTNODE_DIAL_TIMEOUT_SECS: Final = 10.0
-STATUS_REFRESH_INTERVAL_SECS: Final = 1.0
 DEFAULT_NODE_ID: Final = "lean_spec_0"
-PREPARED_ASSETS_DIR: Final = Path(
-    os.environ.get("LEAN_RUNTIME_ASSET_ROOT", "/tmp/lean-spec-client")
-)
-DEVNET3_SOURCE_KEYS_DIR: Final = Path("/app/hive/prod_scheme_devnet3")
-DEVNET4_SOURCE_KEYS_DIR: Final = Path("/app/hive/prod_scheme_devnet4")
-GENESIS_DIR: Final = PREPARED_ASSETS_DIR / "genesis"
-GENESIS_CONFIG_PATH: Final = PREPARED_ASSETS_DIR / "genesis" / "config.yaml"
-VALIDATOR_KEYS_DIR: Final = PREPARED_ASSETS_DIR / "keys"
-MANIFEST_DIR: Final = VALIDATOR_KEYS_DIR / "hash-sig-keys"
-VALIDATORS_YAML_PATH: Final = VALIDATOR_KEYS_DIR / "validators.yaml"
-VALIDATOR_MANIFEST_PATH: Final = (
-    VALIDATOR_KEYS_DIR / "hash-sig-keys" / "validator-keys-manifest.yaml"
-)
 HELPER_IDENTITY_PRIVATE_KEY_HEX: Final = (
     "1111111111111111111111111111111111111111111111111111111111111111"
-)
-SECP256K1_ORDER: Final = int(
-    "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141",
-    16,
 )
 HELPER_IDENTITY_PRIVATE_KEY_ENVIRONMENT_VARIABLE: Final = (
     "HIVE_LEAN_HELPER_IDENTITY_PRIVATE_KEY"
 )
-HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE: Final = "HIVE_LEAN_HELPER_ADVERTISE_IP"
 HELPER_GOSSIP_FORK_DIGEST_ENVIRONMENT_VARIABLE: Final = "HIVE_LEAN_HELPER_GOSSIP_FORK_DIGEST"
 HELPER_P2P_PORT_ENVIRONMENT_VARIABLE: Final = "HIVE_LEAN_HELPER_P2P_PORT"
 HELPER_API_PORT_ENVIRONMENT_VARIABLE: Final = "HIVE_LEAN_HELPER_API_PORT"
@@ -140,349 +100,20 @@ IDENTIFY_PROTOCOL_ID: Final = "/ipfs/id/1.0.0"
 IDENTIFY_PUSH_PROTOCOL_ID: Final = "/ipfs/id/push/1.0.0"
 LIBP2P_SECP256K1_PUBLIC_KEY_TYPE: Final = 2
 DEFAULT_VALIDATOR_COUNT: Final = 3
-ATTESTATION_PUBKEY_FIELD: Final = "attestation_public"
-ATTESTATION_SECRET_FIELD: Final = "attestation_secret"
-PROPOSAL_PUBKEY_FIELD: Final = "proposal_public"
-PROPOSAL_SECRET_FIELD: Final = "proposal_secret"
-LEGACY_PUBLIC_FIELD: Final = "public"
-LEGACY_SECRET_FIELD: Final = "secret"
-WIRE_COMPAT_ENVIRONMENT_VARIABLE: Final = "HIVE_LEAN_SPEC_WIRE_COMPAT"
-WIRE_COMPAT_LEGACY: Final = "legacy"
-WIRE_COMPAT_NATIVE: Final = "native"
 
 logger = logging.getLogger("lean_spec_client_runner")
-LEGACY_SIGNATURE_LENGTH: Final = 2536
-SIGNATURE_PATH_OFFSET: Final = 36
-SIGNATURE_HASHES_OFFSET: Final = 1064
-SIGNATURE_PATH_SIBLINGS_OFFSET: Final = 4
-
-
-def api_server_supports_signed_block_getter() -> bool:
-    return "signed_block_getter" in inspect.signature(ApiServer).parameters
-
-
-def identity_keypair_from_private_key_hex(private_key_hex: str) -> IdentityKeypair:
-    private_key_bytes = Bytes32(bytes.fromhex(private_key_hex))
-
-    if hasattr(IdentityKeypair, "from_bytes"):
-        return IdentityKeypair.from_bytes(private_key_bytes)
-
-    private_key_value = int.from_bytes(private_key_bytes, "big")
-    if not 0 < private_key_value < SECP256K1_ORDER:
-        raise ValueError("helper identity private key is outside the secp256k1 scalar range")
-
-    private_key = ec.derive_private_key(private_key_value, ec.SECP256K1())
-    return IdentityKeypair(
-        private_key=private_key,
-        public_key=Secp256k1PublicKey(_key=private_key.public_key()),
-    )
-
-
-def canonicalize_der_secp256k1_signature(der_signature: bytes) -> bytes:
-    signature_r, signature_s = decode_dss_signature(der_signature)
-    if signature_s > SECP256K1_ORDER // 2:
-        signature_s = SECP256K1_ORDER - signature_s
-    return encode_dss_signature(signature_r, signature_s)
-
-
-def install_low_s_identity_signature_compatibility() -> None:
-    global _LOW_S_IDENTITY_SIGNATURE_COMPAT_INSTALLED
-
-    if _LOW_S_IDENTITY_SIGNATURE_COMPAT_INSTALLED:
-        return
-
-    original_sign = IdentityKeypair.sign
-
-    def sign_low_s(self: IdentityKeypair, message: bytes) -> bytes:
-        return canonicalize_der_secp256k1_signature(original_sign(self, message))
-
-    IdentityKeypair.sign = sign_low_s
-    _LOW_S_IDENTITY_SIGNATURE_COMPAT_INSTALLED = True
 
 
 _QUIC_STREAM_CLOSE_COMPAT_INSTALLED = False
-_SSZ_FAST_DESERIALIZATION_INSTALLED = False
-_SNAPPY_COMPRESS_FALLBACK_INSTALLED = False
 _REQRESP_BLOCK_CACHE_TRACKING_INSTALLED = False
 _NETWORK_SERVICE_GOSSIP_TRACKING_INSTALLED = False
 _IDENTIFY_PROTOCOL_COMPAT_INSTALLED = False
 _CHILD_ONLY_AGGREGATION_COMPAT_INSTALLED = False
-_SETUP_PROVER_COMPAT_INSTALLED = False
 _TRUSTED_GOSSIP_ATTESTATION_COMPAT_INSTALLED = False
 _CHECKPOINT_FINALIZATION_COMPAT_INSTALLED = False
-_LEGACY_SIGNED_BLOCK_WIRE_COMPAT_INSTALLED = False
-_LOW_S_IDENTITY_SIGNATURE_COMPAT_INSTALLED = False
 _BLOCK_CACHE_BY_REQRESP_CLIENT: dict[int, dict[Bytes32, object]] = {}
 _BLOCK_CACHE_BY_SYNC_SERVICE: dict[int, dict[Bytes32, object]] = {}
 _SYNC_EVENT_CONTEXT: dict[int, tuple[LiveNetworkEventSource, Node]] = {}
-
-
-def setup_logging() -> None:
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s %(levelname)-8s %(name)s: %(message)s",
-        datefmt="%Y-%m-%d %H:%M:%S",
-    )
-
-
-def _ssz_offset(value: int) -> bytes:
-    return value.to_bytes(OFFSET_BYTE_LENGTH, "little")
-
-
-def _safe_len(value: object | None) -> int | None:
-    if value is None:
-        return None
-
-    try:
-        return len(value)  # type: ignore[arg-type]
-    except TypeError:
-        return None
-
-
-def legacy_attestation_count(block: object) -> int:
-    body = getattr(block, "body", None)
-    attestations = getattr(body, "attestations", None)
-    if attestations is None:
-        attestations = getattr(block, "attestations", None)
-
-    for candidate in (getattr(attestations, "data", None), attestations):
-        count = _safe_len(candidate)
-        if count is not None:
-            return count
-
-    return 0
-
-
-def _encode_aggregated_signature_proof(
-    participants: bytes = b"\x01",
-    proof_data: bytes = b"",
-) -> bytes:
-    fixed_part_length = OFFSET_BYTE_LENGTH * 2
-    return (
-        _ssz_offset(fixed_part_length)
-        + _ssz_offset(fixed_part_length + len(participants))
-        + participants
-        + proof_data
-    )
-
-
-def _encode_aggregated_signature_proofs(proofs: list[bytes]) -> bytes:
-    offsets = bytearray()
-    offset = OFFSET_BYTE_LENGTH * len(proofs)
-    for proof in proofs:
-        offsets += _ssz_offset(offset)
-        offset += len(proof)
-    return bytes(offsets) + b"".join(proofs)
-
-
-def _encode_placeholder_aggregated_signature_proofs(count: int) -> bytes:
-    return _encode_aggregated_signature_proofs(
-        [_encode_aggregated_signature_proof() for _ in range(count)]
-    )
-
-
-def _encode_blank_xmss_signature() -> bytes:
-    signature = bytearray(LEGACY_SIGNATURE_LENGTH)
-    signature[0:4] = _ssz_offset(SIGNATURE_PATH_OFFSET)
-    signature[32:36] = _ssz_offset(SIGNATURE_HASHES_OFFSET)
-    signature[36:40] = _ssz_offset(SIGNATURE_PATH_SIBLINGS_OFFSET)
-    return bytes(signature)
-
-
-def encode_ssz_bytes(value: object) -> bytes:
-    if isinstance(value, bytes):
-        return value
-    if isinstance(value, bytearray):
-        return bytes(value)
-
-    encode_bytes = getattr(value, "encode_bytes", None)
-    if encode_bytes is not None:
-        return encode_bytes()
-
-    return bytes(value)  # type: ignore[arg-type]
-
-
-def extract_proposer_signature_bytes(signed_block: object) -> bytes:
-    signature = getattr(signed_block, "signature", None)
-    proposer_signature = getattr(signature, "proposer_signature", None)
-    if proposer_signature is None:
-        return _encode_blank_xmss_signature()
-
-    try:
-        encoded = encode_ssz_bytes(proposer_signature)
-    except Exception:
-        logger.debug("Unable to encode proposer signature", exc_info=True)
-        return _encode_blank_xmss_signature()
-
-    if len(encoded) != LEGACY_SIGNATURE_LENGTH:
-        logger.debug("Unexpected proposer signature length %d", len(encoded))
-        return _encode_blank_xmss_signature()
-
-    return encoded
-
-
-def extract_attestation_signature_proofs(signed_block: object) -> list[bytes]:
-    signature = getattr(signed_block, "signature", None)
-    attestation_signatures = getattr(signature, "attestation_signatures", None)
-    proof_values = getattr(attestation_signatures, "data", attestation_signatures)
-    if proof_values is None:
-        return []
-
-    proofs = []
-    for proof in proof_values:
-        try:
-            participants = encode_ssz_bytes(getattr(proof, "participants"))
-            proof_data = encode_ssz_bytes(getattr(proof, "proof_data"))
-            proofs.append(_encode_aggregated_signature_proof(participants, proof_data))
-        except Exception:
-            logger.debug("Unable to encode attestation signature proof", exc_info=True)
-            proofs.append(_encode_aggregated_signature_proof())
-
-    return proofs
-
-
-def _encode_legacy_block_signatures(
-    proposer_signature: bytes,
-    attestation_signature_proofs: bytes,
-) -> bytes:
-    # Checkpoint clients validate that this list has one placeholder proof per
-    # attestation, but they do not need the proofs to be valid for bootstrapping.
-    fixed_part_length = OFFSET_BYTE_LENGTH + LEGACY_SIGNATURE_LENGTH
-    return (
-        _ssz_offset(fixed_part_length)
-        + proposer_signature
-        + attestation_signature_proofs
-    )
-
-
-def _encode_legacy_signed_block(
-    block_bytes: bytes,
-    proposer_signature: bytes,
-    attestation_signature_proofs: bytes,
-) -> bytes:
-    signatures_bytes = _encode_legacy_block_signatures(
-        proposer_signature,
-        attestation_signature_proofs,
-    )
-    fixed_part_length = OFFSET_BYTE_LENGTH * 2
-    return (
-        _ssz_offset(fixed_part_length)
-        + _ssz_offset(fixed_part_length + len(block_bytes))
-        + block_bytes
-        + signatures_bytes
-    )
-
-
-def encode_legacy_signed_block(signed_block: object) -> bytes:
-    block = extract_inner_block(signed_block)
-    proofs = extract_attestation_signature_proofs(signed_block)
-    if not proofs:
-        proof_bytes = _encode_placeholder_aggregated_signature_proofs(
-            legacy_attestation_count(block)
-        )
-    else:
-        proof_bytes = _encode_aggregated_signature_proofs(proofs)
-    return _encode_legacy_signed_block(
-        block.encode_bytes(),
-        extract_proposer_signature_bytes(signed_block),
-        proof_bytes,
-    )
-
-def install_legacy_signed_block_wire_compatibility() -> None:
-    global _LEGACY_SIGNED_BLOCK_WIRE_COMPAT_INSTALLED
-
-    if _LEGACY_SIGNED_BLOCK_WIRE_COMPAT_INSTALLED:
-        return
-
-    if not legacy_signed_block_compatibility_enabled():
-        _LEGACY_SIGNED_BLOCK_WIRE_COMPAT_INSTALLED = True
-        return
-
-    networking_service_module = importlib.import_module(
-        "lean_spec.subspecs.networking.service.service"
-    )
-    reqresp_handler_module = importlib.import_module(
-        "lean_spec.subspecs.networking.reqresp.handler"
-    )
-
-    network_service_cls = networking_service_module.NetworkService
-    request_handler_cls = reqresp_handler_module.RequestHandler
-    response_code_cls = reqresp_handler_module.ResponseCode
-    slot_cls = reqresp_handler_module.Slot
-    uint64_cls = reqresp_handler_module.Uint64
-
-    async def publish_block_with_legacy_signed_block(self: object, block: object) -> None:
-        topic = networking_service_module.GossipTopic.block(self.network_name)
-        compressed = networking_service_module.compress(encode_legacy_signed_block(block))
-        await self.event_source.publish(topic.to_topic_id(), compressed)
-        logger.debug("Published legacy signed block at slot %s", extract_inner_block(block).slot)
-
-    async def handle_blocks_by_root_with_legacy_signed_block(
-        self: object,
-        request: object,
-        response: object,
-    ) -> None:
-        if self.block_lookup is None:
-            logger.warning("BlocksByRoot request received but no block_lookup configured")
-            await response.send_error(response_code_cls.SERVER_ERROR, "Block lookup not available")
-            return
-
-        for root in request.roots.data:
-            try:
-                block = await self.block_lookup(root)
-                if block is not None:
-                    await response.send_success(encode_legacy_signed_block(block))
-            except Exception as err:
-                logger.warning("Error looking up block %s: %s", root.hex()[:8], err)
-
-    async def handle_blocks_by_range_with_legacy_signed_block(
-        self: object,
-        request: object,
-        response: object,
-    ) -> None:
-        if self.block_by_slot_lookup is None:
-            logger.warning("BlocksByRange request received but no block_by_slot_lookup configured")
-            await response.send_error(response_code_cls.SERVER_ERROR, "Block lookup not available")
-            return
-
-        if request.count == uint64_cls(0) or request.count > uint64_cls(
-            reqresp_handler_module.MAX_REQUEST_BLOCKS
-        ):
-            await response.send_error(response_code_cls.INVALID_REQUEST, "Invalid count")
-            return
-
-        if self.current_slot_lookup is None:
-            logger.warning("BlocksByRange request received but no current_slot_lookup configured")
-            await response.send_error(response_code_cls.SERVER_ERROR, "Current slot not available")
-            return
-
-        current_slot = self.current_slot_lookup()
-        window_floor = (
-            current_slot - slot_cls(reqresp_handler_module.MIN_SLOTS_FOR_BLOCK_REQUESTS)
-            if current_slot >= slot_cls(reqresp_handler_module.MIN_SLOTS_FOR_BLOCK_REQUESTS)
-            else slot_cls(0)
-        )
-        if request.start_slot < window_floor:
-            await response.send_error(
-                response_code_cls.RESOURCE_UNAVAILABLE,
-                "Requested slot predates history window",
-            )
-            return
-
-        for index in range(int(request.count)):
-            slot = request.start_slot + slot_cls(index)
-            try:
-                block = await self.block_by_slot_lookup(slot)
-                if block is not None:
-                    await response.send_success(encode_legacy_signed_block(block))
-            except Exception as err:
-                logger.warning("Error looking up block at slot %s: %s", slot, err)
-
-    network_service_cls.publish_block = publish_block_with_legacy_signed_block
-    request_handler_cls.handle_blocks_by_root = handle_blocks_by_root_with_legacy_signed_block
-    request_handler_cls.handle_blocks_by_range = handle_blocks_by_range_with_legacy_signed_block
-    logger.info("Installed LeanSpec signed-block gossip/reqresp compatibility")
-    _LEGACY_SIGNED_BLOCK_WIRE_COMPAT_INSTALLED = True
 
 
 def install_quic_stream_close_reset_tolerance() -> None:
@@ -510,209 +141,6 @@ def install_quic_stream_close_reset_tolerance() -> None:
 
     QuicStream.close = close_ignoring_reset
     _QUIC_STREAM_CLOSE_COMPAT_INSTALLED = True
-
-
-def install_fast_ssz_deserialization() -> None:
-    global _SSZ_FAST_DESERIALIZATION_INSTALLED
-
-    if _SSZ_FAST_DESERIALIZATION_INSTALLED:
-        return
-
-    @classmethod
-    def vector_deserialize_without_validation(
-        cls: type[SSZVector], stream: io.BytesIO, scope: int
-    ) -> SSZVector:
-        elements = []
-        if cls.is_fixed_size():
-            elem_byte_length = cls.get_byte_length() // cls.LENGTH
-            if scope != cls.get_byte_length():
-                raise SSZSerializationError(
-                    f"{cls.__name__}: expected {cls.get_byte_length()} bytes, got {scope}"
-                )
-            elements = [
-                cls.ELEMENT_TYPE.deserialize(stream, elem_byte_length)
-                for _ in range(cls.LENGTH)
-            ]
-        else:
-            first_offset = int(Uint32.deserialize(stream, OFFSET_BYTE_LENGTH))
-            expected = cls.LENGTH * OFFSET_BYTE_LENGTH
-            if first_offset != expected:
-                raise SSZSerializationError(
-                    f"{cls.__name__}: invalid offset {first_offset}, expected {expected}"
-                )
-            offsets = [first_offset] + [
-                int(Uint32.deserialize(stream, OFFSET_BYTE_LENGTH))
-                for _ in range(cls.LENGTH - 1)
-            ]
-            offsets.append(scope)
-            _validate_offsets(offsets, scope, cls.__name__)
-            for i in range(cls.LENGTH):
-                start, end = offsets[i], offsets[i + 1]
-                elements.append(cls.ELEMENT_TYPE.deserialize(stream, end - start))
-
-        return cls.model_construct(data=tuple(elements))
-
-    @classmethod
-    def list_deserialize_without_validation(
-        cls: type[SSZList], stream: io.BytesIO, scope: int
-    ) -> SSZList:
-        if cls.ELEMENT_TYPE.is_fixed_size():
-            element_size = cls.ELEMENT_TYPE.get_byte_length()
-            if scope % element_size != 0:
-                raise SSZSerializationError(
-                    f"{cls.__name__}: scope {scope} not divisible by element size {element_size}"
-                )
-
-            num_elements = scope // element_size
-            if num_elements > cls.LIMIT:
-                raise SSZValueError(
-                    f"{cls.__name__} exceeds limit of {cls.LIMIT}, got {num_elements}"
-                )
-
-            elements = [
-                cls.ELEMENT_TYPE.deserialize(stream, element_size)
-                for _ in range(num_elements)
-            ]
-            return cls.model_construct(data=tuple(elements))
-
-        if scope == 0:
-            return cls.model_construct(data=tuple())
-        if scope < OFFSET_BYTE_LENGTH:
-            raise SSZSerializationError(
-                f"{cls.__name__}: scope {scope} too small for variable-size list"
-            )
-
-        first_offset = int(Uint32.deserialize(stream, OFFSET_BYTE_LENGTH))
-        if first_offset > scope or first_offset % OFFSET_BYTE_LENGTH != 0:
-            raise SSZSerializationError(f"{cls.__name__}: invalid offset {first_offset}")
-
-        count = first_offset // OFFSET_BYTE_LENGTH
-        if count > cls.LIMIT:
-            raise SSZValueError(f"{cls.__name__} exceeds limit of {cls.LIMIT}, got {count}")
-
-        offsets = [first_offset] + [
-            int(Uint32.deserialize(stream, OFFSET_BYTE_LENGTH)) for _ in range(count - 1)
-        ]
-        offsets.append(scope)
-        _validate_offsets(offsets, scope, cls.__name__)
-
-        elements = []
-        for i in range(count):
-            start, end = offsets[i], offsets[i + 1]
-            elements.append(cls.ELEMENT_TYPE.deserialize(stream, end - start))
-
-        return cls.model_construct(data=tuple(elements))
-
-    @classmethod
-    def container_deserialize_without_validation(
-        cls: type[Container], stream: io.BytesIO, scope: int
-    ) -> Container:
-        fields: dict[str, object] = {}
-        var_fields = []
-        bytes_read = 0
-
-        for field_name, field_info in cls.model_fields.items():
-            field_type = cls._get_ssz_field_type(field_info.annotation)
-
-            if field_type.is_fixed_size():
-                size = field_type.get_byte_length()
-                data = stream.read(size)
-                if len(data) != size:
-                    raise SSZSerializationError(
-                        f"{cls.__name__}.{field_name}: expected {size} bytes, got {len(data)}"
-                    )
-                fields[field_name] = field_type.decode_bytes(data)
-                bytes_read += size
-            else:
-                offset_bytes = stream.read(OFFSET_BYTE_LENGTH)
-                if len(offset_bytes) != OFFSET_BYTE_LENGTH:
-                    raise SSZSerializationError(
-                        f"{cls.__name__}.{field_name}: "
-                        f"expected {OFFSET_BYTE_LENGTH} offset bytes, got {len(offset_bytes)}"
-                    )
-                offset = int(Uint32.decode_bytes(offset_bytes))
-                var_fields.append((field_name, field_type, offset))
-                bytes_read += OFFSET_BYTE_LENGTH
-
-        if var_fields:
-            var_section_size = scope - bytes_read
-            var_section = stream.read(var_section_size)
-            if len(var_section) != var_section_size:
-                raise SSZSerializationError(
-                    f"{cls.__name__}: "
-                    f"expected {var_section_size} variable bytes, got {len(var_section)}"
-                )
-
-            offsets = [offset for _, _, offset in var_fields] + [scope]
-            for i, (name, field_type, start) in enumerate(var_fields):
-                end = offsets[i + 1]
-                rel_start = start - bytes_read
-                rel_end = end - bytes_read
-                if rel_start < 0 or rel_start > rel_end:
-                    raise SSZSerializationError(
-                        f"{cls.__name__}.{name}: invalid offsets start={start}, end={end}"
-                    )
-                fields[name] = field_type.decode_bytes(var_section[rel_start:rel_end])
-
-        return cls.model_construct(**fields)
-
-    SSZVector.deserialize = vector_deserialize_without_validation
-    SSZList.deserialize = list_deserialize_without_validation
-    Container.deserialize = container_deserialize_without_validation
-    _SSZ_FAST_DESERIALIZATION_INSTALLED = True
-
-
-def install_snappy_compress_fallback() -> None:
-    global _SNAPPY_COMPRESS_FALLBACK_INSTALLED
-
-    if _SNAPPY_COMPRESS_FALLBACK_INSTALLED:
-        return
-
-    snappy_package_module = importlib.import_module("lean_spec.snappy")
-    snappy_compress_module = importlib.import_module("lean_spec.snappy.compress")
-    snappy_framing_module = importlib.import_module("lean_spec.snappy.framing")
-    reqresp_codec_module = importlib.import_module("lean_spec.subspecs.networking.reqresp.codec")
-    networking_service_module = importlib.import_module(
-        "lean_spec.subspecs.networking.service.service"
-    )
-
-    def literal_only_compress(data: bytes) -> bytes:
-        if not data:
-            return snappy_compress_module.encode_varint32(0)
-
-        output = bytearray(snappy_compress_module.encode_varint32(len(data)))
-        offset = 0
-        while offset < len(data):
-            block_end = min(offset + snappy_compress_module.BLOCK_SIZE, len(data))
-            output.extend(snappy_compress_module._emit_literal(data[offset:block_end]))
-            offset = block_end
-
-        return bytes(output)
-
-    def compress_with_fallback(data: bytes) -> bytes:
-        return literal_only_compress(data)
-
-    def frame_compress_with_fallback(data: bytes) -> bytes:
-        output = bytearray(snappy_framing_module.STREAM_IDENTIFIER)
-        offset = 0
-        while offset < len(data):
-            chunk_end = min(offset + snappy_framing_module.MAX_UNCOMPRESSED_CHUNK_SIZE, len(data))
-            chunk = data[offset:chunk_end]
-            offset = chunk_end
-            crc = snappy_framing_module._mask_crc(snappy_framing_module._crc32c(chunk))
-            output.append(snappy_framing_module.CHUNK_TYPE_UNCOMPRESSED)
-            output.extend((4 + len(chunk)).to_bytes(3, "little"))
-            output.extend(crc.to_bytes(4, "little"))
-            output.extend(chunk)
-        return bytes(output)
-
-    snappy_compress_module.compress = compress_with_fallback
-    snappy_framing_module.frame_compress = frame_compress_with_fallback
-    snappy_package_module.compress = compress_with_fallback
-    snappy_package_module.frame_compress = frame_compress_with_fallback
-    reqresp_codec_module.frame_compress = frame_compress_with_fallback
-    networking_service_module.compress = compress_with_fallback
-    _SNAPPY_COMPRESS_FALLBACK_INSTALLED = True
 
 
 def install_reqresp_block_cache_tracking() -> None:
@@ -1019,51 +447,6 @@ def install_child_only_aggregation_compatibility() -> None:
     _CHILD_ONLY_AGGREGATION_COMPAT_INSTALLED = True
 
 
-def install_setup_prover_compatibility() -> None:
-    global _SETUP_PROVER_COMPAT_INSTALLED
-
-    if _SETUP_PROVER_COMPAT_INSTALLED:
-        return
-
-    original_setup_prover = getattr(xmss_aggregation_module, "setup_prover", None)
-    original_setup_verifier = getattr(xmss_aggregation_module, "setup_verifier", None)
-    if original_setup_prover is None and original_setup_verifier is None:
-        _SETUP_PROVER_COMPAT_INSTALLED = True
-        return
-
-    initialized_prover_modes: set[str] = set()
-    initialized_verifier_modes: set[str] = set()
-    default_mode = getattr(xmss_aggregation_module, "LEAN_ENV", os.environ.get("LEAN_ENV"))
-
-    def setup_once(
-        original_setup: object | None,
-        initialized_modes: set[str],
-        mode: object | None,
-    ) -> None:
-        if original_setup is None:
-            return
-        key = str(mode)
-        if key in initialized_modes:
-            return
-        if mode is None:
-            original_setup()
-        else:
-            original_setup(mode=mode)
-        initialized_modes.add(key)
-
-    def setup_prover_once(*, mode: object | None = None) -> None:
-        setup_once(original_setup_prover, initialized_prover_modes, mode or default_mode)
-
-    def setup_verifier_once(*, mode: object | None = None) -> None:
-        setup_once(original_setup_verifier, initialized_verifier_modes, mode or default_mode)
-
-    if original_setup_prover is not None:
-        xmss_aggregation_module.setup_prover = setup_prover_once
-    if original_setup_verifier is not None:
-        xmss_aggregation_module.setup_verifier = setup_verifier_once
-    _SETUP_PROVER_COMPAT_INSTALLED = True
-
-
 def install_trusted_gossip_attestation_compatibility() -> None:
     global _TRUSTED_GOSSIP_ATTESTATION_COMPAT_INSTALLED
 
@@ -1257,22 +640,6 @@ def install_checkpoint_finalization_compatibility() -> None:
     logger.info("Installed LeanSpec checkpoint finalization compatibility")
 
 
-def parse_bootnodes() -> list[str]:
-    raw_bootnodes = os.environ.get("HIVE_BOOTNODES", "")
-    if not raw_bootnodes:
-        return []
-
-    return [bootnode for bootnode in raw_bootnodes.split(",") if bootnode]
-
-
-def parse_validator_indices() -> list[int]:
-    raw_indices = os.environ.get("HIVE_LEAN_VALIDATOR_INDICES", "")
-    if not raw_indices:
-        return []
-
-    return [int(index) for index in raw_indices.split(",") if index]
-
-
 def genesis_validator_count() -> int:
     raw_count = os.environ.get(GENESIS_VALIDATOR_COUNT_ENVIRONMENT_VARIABLE, "")
     if not raw_count:
@@ -1282,10 +649,6 @@ def genesis_validator_count() -> int:
     if count < 0:
         raise ValueError(f"{GENESIS_VALIDATOR_COUNT_ENVIRONMENT_VARIABLE} must be non-negative")
     return count
-
-
-def devnet_label() -> str:
-    return os.environ.get("HIVE_LEAN_DEVNET_LABEL", "devnet3")
 
 
 def helper_gossip_fork_digest() -> str:
@@ -1316,268 +679,6 @@ def listen_addr() -> str:
     return f"/ip4/0.0.0.0/udp/{helper_p2p_port()}/quic-v1"
 
 
-def uses_latest_leanspec_format() -> bool:
-    return devnet_label() == "devnet4"
-
-
-def wire_compat_mode() -> str:
-    mode = os.environ.get(WIRE_COMPAT_ENVIRONMENT_VARIABLE, WIRE_COMPAT_NATIVE).strip().lower()
-    if mode not in {WIRE_COMPAT_LEGACY, WIRE_COMPAT_NATIVE}:
-        logger.warning(
-            "Unknown %s=%r; falling back to %s",
-            WIRE_COMPAT_ENVIRONMENT_VARIABLE,
-            mode,
-            WIRE_COMPAT_NATIVE,
-        )
-        return WIRE_COMPAT_NATIVE
-    return mode
-
-
-def legacy_signed_block_compatibility_enabled() -> bool:
-    return uses_latest_leanspec_format() and wire_compat_mode() == WIRE_COMPAT_LEGACY
-
-
-def load_validator(index: int) -> dict[str, str]:
-    with (source_keys_dir() / f"{index}.json").open(encoding="utf-8") as key_file:
-        validator = json.load(key_file)
-
-    if {"attestation_keypair", "proposal_keypair"}.issubset(validator):
-        return {
-            ATTESTATION_PUBKEY_FIELD: validator["attestation_keypair"]["public_key"],
-            ATTESTATION_SECRET_FIELD: validator["attestation_keypair"]["secret_key"],
-            PROPOSAL_PUBKEY_FIELD: validator["proposal_keypair"]["public_key"],
-            PROPOSAL_SECRET_FIELD: validator["proposal_keypair"]["secret_key"],
-        }
-
-    if {
-        ATTESTATION_PUBKEY_FIELD,
-        ATTESTATION_SECRET_FIELD,
-        PROPOSAL_PUBKEY_FIELD,
-        PROPOSAL_SECRET_FIELD,
-    }.issubset(validator):
-        return validator
-
-    if {LEGACY_PUBLIC_FIELD, LEGACY_SECRET_FIELD}.issubset(validator):
-        # Latest published devnet4 helper bundles currently expose a single
-        # XMSS keypair per validator. LeanSpec still expects separate
-        # attestation/proposal entries, so we mirror the same key material
-        # into both roles for test-only helper startup compatibility.
-        return {
-            ATTESTATION_PUBKEY_FIELD: validator[LEGACY_PUBLIC_FIELD],
-            ATTESTATION_SECRET_FIELD: validator[LEGACY_SECRET_FIELD],
-            PROPOSAL_PUBKEY_FIELD: validator[LEGACY_PUBLIC_FIELD],
-            PROPOSAL_SECRET_FIELD: validator[LEGACY_SECRET_FIELD],
-        }
-
-    raise ValueError(
-        f"Unsupported validator key format for validator {index}: "
-        f"expected nested keypairs, split attestation/proposal fields, or "
-        f"{LEGACY_PUBLIC_FIELD!r}/{LEGACY_SECRET_FIELD!r}"
-    )
-
-
-def decode_secret_key(secret_key_hex: str) -> SecretKey:
-    # Current devnet4 prod keys are large enough that LeanSpec's pydantic-heavy
-    # decode path can trip GC while nested objects are still being built.
-    was_gc_enabled = gc.isenabled()
-    if was_gc_enabled:
-        gc.disable()
-    try:
-        return SecretKey.decode_bytes(bytes.fromhex(secret_key_hex))
-    finally:
-        if was_gc_enabled:
-            gc.enable()
-
-
-def source_keys_dir() -> Path:
-    if uses_latest_leanspec_format():
-        return DEVNET4_SOURCE_KEYS_DIR
-
-    return DEVNET3_SOURCE_KEYS_DIR
-
-
-def prepare_output_dirs() -> None:
-    if PREPARED_ASSETS_DIR.exists():
-        shutil.rmtree(PREPARED_ASSETS_DIR)
-
-    GENESIS_DIR.mkdir(parents=True, exist_ok=True)
-    MANIFEST_DIR.mkdir(parents=True, exist_ok=True)
-
-
-def write_genesis(validators: list[dict[str, str]], genesis_time: int) -> None:
-    genesis = {
-        "GENESIS_TIME": genesis_time,
-        "ATTESTATION_COMMITTEE_COUNT": 1,
-        "MAX_ATTESTATIONS_DATA": 16,
-        "NUM_VALIDATORS": len(validators),
-        "VALIDATOR_COUNT": len(validators),
-        "ACTIVE_EPOCH": 18,
-    }
-
-    if uses_latest_leanspec_format():
-        genesis["GENESIS_VALIDATORS"] = [
-            {
-                "attestation_pubkey": f"0x{validator[ATTESTATION_PUBKEY_FIELD]}",
-                "proposal_pubkey": f"0x{validator[PROPOSAL_PUBKEY_FIELD]}",
-            }
-            for validator in validators
-        ]
-    else:
-        genesis["GENESIS_VALIDATORS"] = [
-            f"0x{validator[ATTESTATION_PUBKEY_FIELD]}" for validator in validators
-        ]
-
-    with GENESIS_CONFIG_PATH.open("w", encoding="utf-8") as genesis_file:
-        yaml.safe_dump(genesis, genesis_file, sort_keys=False)
-
-
-def write_validator_keys(
-    validators: list[dict[str, str]], node_id: str, validator_indices: list[int]
-) -> None:
-    with VALIDATORS_YAML_PATH.open("w", encoding="utf-8") as validators_file:
-        yaml.safe_dump({node_id: validator_indices}, validators_file, sort_keys=False)
-
-    manifest = {
-        "key_scheme": "SIGTopLevelTargetSumLifetime32Dim64Base8",
-        "hash_function": "Poseidon2",
-        "encoding": "TargetSum",
-        "lifetime": 32,
-        "log_num_active_epochs": 5,
-        "num_active_epochs": 32,
-        "num_validators": len(validators),
-        "validators": [],
-    }
-
-    for index, validator in enumerate(validators):
-        if uses_latest_leanspec_format():
-            attestation_secret_key_file = f"validator_{index}_att_sk.ssz"
-            proposal_secret_key_file = f"validator_{index}_prop_sk.ssz"
-
-            (MANIFEST_DIR / attestation_secret_key_file).write_bytes(
-                bytes.fromhex(validator[ATTESTATION_SECRET_FIELD])
-            )
-            (MANIFEST_DIR / proposal_secret_key_file).write_bytes(
-                bytes.fromhex(validator[PROPOSAL_SECRET_FIELD])
-            )
-
-            manifest["validators"].append(
-                {
-                    "index": index,
-                    "attestation_pubkey_hex": f"0x{validator[ATTESTATION_PUBKEY_FIELD]}",
-                    "proposal_pubkey_hex": f"0x{validator[PROPOSAL_PUBKEY_FIELD]}",
-                    "attestation_privkey_file": attestation_secret_key_file,
-                    "proposal_privkey_file": proposal_secret_key_file,
-                }
-            )
-        else:
-            secret_key_file = f"validator_{index}_sk.ssz"
-
-            (MANIFEST_DIR / secret_key_file).write_bytes(
-                bytes.fromhex(validator[ATTESTATION_SECRET_FIELD])
-            )
-
-            manifest["validators"].append(
-                {
-                    "index": index,
-                    "pubkey_hex": f"0x{validator[ATTESTATION_PUBKEY_FIELD]}",
-                    "privkey_file": secret_key_file,
-                }
-            )
-
-    with VALIDATOR_MANIFEST_PATH.open("w", encoding="utf-8") as manifest_file:
-        yaml.safe_dump(manifest, manifest_file, sort_keys=False)
-
-
-def prepare_runtime_assets(node_id: str) -> None:
-    genesis_time = int(os.environ.get("HIVE_LEAN_GENESIS_TIME", int(time.time()) + 30))
-    validator_indices = parse_validator_indices()
-    validator_count = genesis_validator_count()
-    if validator_indices and max(validator_indices) >= validator_count:
-        raise ValueError(
-            f"{GENESIS_VALIDATOR_COUNT_ENVIRONMENT_VARIABLE}={validator_count} does not include "
-            f"assigned validator index {max(validator_indices)}"
-        )
-    validators = [load_validator(index) for index in range(validator_count)]
-
-    prepare_output_dirs()
-    write_genesis(validators, genesis_time)
-    write_validator_keys(validators, node_id, validator_indices)
-
-
-def helper_genesis_metadata() -> dict[str, object]:
-    with GENESIS_CONFIG_PATH.open(encoding="utf-8") as genesis_file:
-        genesis_config = yaml.safe_load(genesis_file)
-
-    validators = genesis_config.get("GENESIS_VALIDATORS", [])
-    if uses_latest_leanspec_format():
-        genesis_validator_entries = [
-            {
-                "attestation_public_key": validator["attestation_pubkey"],
-                "proposal_public_key": validator["proposal_pubkey"],
-            }
-            for validator in validators
-        ]
-    else:
-        genesis_validator_entries = [
-            {
-                "attestation_public_key": validator,
-                "proposal_public_key": None,
-            }
-            for validator in validators
-        ]
-
-    return {
-        "genesis_time": genesis_config["GENESIS_TIME"],
-        "genesis_validator_entries": genesis_validator_entries,
-    }
-
-
-def encode_enr_port(port: int) -> bytes:
-    byte_length = max(1, (port.bit_length() + 7) // 8)
-    return port.to_bytes(byte_length, "big")
-
-
-def build_canonical_secp256k1_signature(
-    identity_key: IdentityKeypair,
-    digest: bytes,
-) -> bytes:
-    der_signature = identity_key.private_key.sign(
-        digest,
-        ec.ECDSA(Prehashed(hashes.SHA256())),
-    )
-    signature_r, signature_s = decode_dss_signature(der_signature)
-    if signature_s > SECP256K1_ORDER // 2:
-        signature_s = SECP256K1_ORDER - signature_s
-    return signature_r.to_bytes(32, "big") + signature_s.to_bytes(32, "big")
-
-
-def build_helper_bootnode_enr(
-    identity_key: IdentityKeypair,
-    *,
-    include_udp: bool = True,
-) -> str:
-    advertise_ip = os.environ.get(HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE)
-    if not advertise_ip:
-        raise ValueError(
-            f"Missing {HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE} for helper ENR generation"
-        )
-
-    pairs = {
-        enr_keys.ID: b"v4",
-        enr_keys.SECP256K1: identity_key.public_key.to_bytes(),
-        enr_keys.IP: ipaddress.ip_address(advertise_ip).packed,
-    }
-    if include_udp:
-        pairs[enr_keys.UDP] = encode_enr_port(helper_p2p_port())
-    pairs[enr_keys.QUIC] = encode_enr_port(helper_p2p_port())
-
-    unsigned_enr = ENR(signature=b"\x00" * 64, seq=1, pairs=pairs)
-    digest = keccak.new(digest_bits=256, data=unsigned_enr._content_rlp()).digest()
-    # Normalize helper ECDSA signatures so every client sees a canonical ENR.
-    signature = build_canonical_secp256k1_signature(identity_key, digest)
-    return ENR(signature=signature, seq=1, pairs=pairs).to_string()
-
-
 def build_helper_bootnode_multiaddr(peer_id: str) -> str:
     advertise_ip = os.environ.get(HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE)
     if not advertise_ip:
@@ -1586,45 +687,6 @@ def build_helper_bootnode_multiaddr(peer_id: str) -> str:
         )
 
     return f"/ip4/{advertise_ip}/udp/{helper_p2p_port()}/quic-v1/p2p/{peer_id}"
-
-
-async def start_metadata_server(metadata: dict[str, object]) -> web.AppRunner:
-    async def metadata_handler(_request: web.Request) -> web.Response:
-        return web.json_response(metadata)
-
-    app = web.Application()
-    app.router.add_get("/hive/genesis", metadata_handler)
-
-    runner = web.AppRunner(app)
-    await runner.setup()
-    site = web.TCPSite(runner, host="0.0.0.0", port=helper_metadata_port())
-    await site.start()
-    logger.info("Helper metadata server listening on 0.0.0.0:%d", helper_metadata_port())
-    return runner
-
-
-def resolve_bootnode(bootnode: str) -> str:
-    if bootnode.startswith("enr:"):
-        enr = ENR.from_string(bootnode)
-        if not enr.is_valid():
-            raise ValueError(f"ENR structurally invalid: {bootnode}")
-        if not enr.verify_signature():
-            raise ValueError(f"ENR signature verification failed: {bootnode}")
-
-        multiaddr = enr.multiaddr()
-        if multiaddr is None:
-            raise ValueError(f"ENR has no UDP connection info: {bootnode}")
-        return multiaddr
-
-    return bootnode
-
-
-def load_genesis() -> GenesisConfig:
-    if not GENESIS_CONFIG_PATH.exists():
-        raise FileNotFoundError(
-            f"Prepared LeanSpec genesis config is missing: {GENESIS_CONFIG_PATH}"
-        )
-    return GenesisConfig.from_yaml_file(GENESIS_CONFIG_PATH)
 
 
 def load_validator_registry(node_id: str) -> ValidatorRegistry | None:
@@ -1692,140 +754,7 @@ def load_validator_registry(node_id: str) -> ValidatorRegistry | None:
             if was_gc_enabled:
                 gc.enable()
 
-    if not VALIDATORS_YAML_PATH.exists() or not VALIDATOR_MANIFEST_PATH.exists():
-        return None
-
-    registry = ValidatorRegistry.from_yaml(
-        node_id=node_id,
-        validators_path=VALIDATORS_YAML_PATH,
-        manifest_path=VALIDATOR_MANIFEST_PATH,
-    )
-    return registry if len(registry) > 0 else None
-
-
-def subscribe_gossip_topics(
-    event_source: LiveNetworkEventSource,
-    validator_registry: ValidatorRegistry | None,
-    is_aggregator: bool,
-) -> None:
-    fork_digest = helper_gossip_fork_digest()
-    block_topic = GossipTopic.block(fork_digest).to_topic_id()
-    event_source.subscribe_gossip_topic(block_topic)
-
-    if is_aggregator and hasattr(GossipTopic, "committee_aggregation"):
-        aggregation_topic = GossipTopic.committee_aggregation(fork_digest).to_topic_id()
-        event_source.subscribe_gossip_topic(aggregation_topic)
-
-    subscribed_subnets: set[SubnetId] = set()
-    if validator_registry is not None:
-        for validator_index in validator_registry.indices():
-            subscribed_subnets.add(
-                validator_index.compute_subnet_id(ATTESTATION_COMMITTEE_COUNT)
-            )
-
-    if is_aggregator and not subscribed_subnets:
-        subscribed_subnets.add(SubnetId(0))
-
-    for subnet_id in subscribed_subnets:
-        topic = GossipTopic.attestation_subnet(fork_digest, subnet_id).to_topic_id()
-        event_source.subscribe_gossip_topic(topic)
-
-
-def current_status(node: Node) -> Status:
-    store = node.sync_service.store
-    head_root = store.head
-    head_block = store.blocks[head_root]
-    return Status(
-        finalized=store.latest_finalized,
-        head=Checkpoint(root=head_root, slot=head_block.slot),
-    )
-
-
-def refresh_status(event_source: LiveNetworkEventSource, node: Node) -> None:
-    event_source.set_status(current_status(node))
-
-
-def extract_inner_block(signed_block: object) -> object:
-    if hasattr(signed_block, "block"):
-        return getattr(signed_block, "block")
-
-    if hasattr(signed_block, "slot") and hasattr(signed_block, "parent_root"):
-        return signed_block
-
-    message = getattr(signed_block, "message", None)
-    if message is not None:
-        if hasattr(message, "block"):
-            return message.block
-
-        # Latest LeanSpec SignedBlock wraps the Block directly in `.message`.
-        if hasattr(message, "slot") and hasattr(message, "parent_root"):
-            return message
-
-    raise AttributeError(
-        "Signed block has neither `.block`, `.message.block`, nor a block-like `.message`"
-    )
-
-
-def cache_signed_block(
-    block_cache: dict[Bytes32, object],
-    signed_block: object,
-) -> Bytes32:
-    block_root = hash_tree_root(extract_inner_block(signed_block))
-    block_cache[block_root] = signed_block
-    return block_root
-
-
-async def keep_status_current(
-    event_source: LiveNetworkEventSource,
-    node: Node,
-) -> None:
-    while True:
-        refresh_status(event_source, node)
-        await asyncio.sleep(STATUS_REFRESH_INTERVAL_SECS)
-
-
-async def start_listener_and_gossipsub(
-    event_source: LiveNetworkEventSource,
-) -> asyncio.Task[None]:
-    logger.info("Starting listener on %s", listen_addr())
-    listener_task = asyncio.create_task(
-        event_source.listen(listen_addr()),
-        name="lean-spec-listener",
-    )
-    await asyncio.sleep(0.1)
-
-    if listener_task.done():
-        listener_task.result()
-
-    logger.info("Starting gossipsub behavior...")
-    await event_source.start_gossipsub()
-    return listener_task
-
-
-async def dial_bootnodes(
-    event_source: LiveNetworkEventSource,
-    bootnodes: list[str],
-) -> None:
-    for bootnode in bootnodes:
-        multiaddr = resolve_bootnode(bootnode)
-        logger.info("Connecting to bootnode %s", multiaddr)
-        try:
-            peer_id = await asyncio.wait_for(
-                event_source.dial(multiaddr),
-                timeout=BOOTNODE_DIAL_TIMEOUT_SECS,
-            )
-            if peer_id is None:
-                logger.warning("Failed to connect to bootnode %s", multiaddr)
-            else:
-                logger.info("Connected to bootnode, peer_id=%s", peer_id)
-        except asyncio.TimeoutError:
-            logger.warning(
-                "Timed out after %.1fs while connecting to bootnode %s",
-                BOOTNODE_DIAL_TIMEOUT_SECS,
-                multiaddr,
-            )
-        except Exception:
-            logger.warning("Failed to connect to bootnode %s", multiaddr, exc_info=True)
+    return load_validator_registry_from_manifest(ValidatorRegistry, node_id)
 
 
 async def run() -> None:
@@ -1847,7 +776,19 @@ async def run() -> None:
     metrics.init(name="lean-spec-client", version="0.0.1")
 
     node_id = os.environ.get("HIVE_NODE_ID", DEFAULT_NODE_ID)
-    prepare_runtime_assets(node_id)
+    validator_count = genesis_validator_count()
+    prepare_runtime_assets(
+        node_id,
+        validator_count,
+        genesis_extra_fields={
+            "ATTESTATION_COMMITTEE_COUNT": 1,
+            "MAX_ATTESTATIONS_DATA": 16,
+            "NUM_VALIDATORS": validator_count,
+            "VALIDATOR_COUNT": validator_count,
+            "ACTIVE_EPOCH": 18,
+        },
+        validator_count_environment_variable=GENESIS_VALIDATOR_COUNT_ENVIRONMENT_VARIABLE,
+    )
     genesis = load_genesis()
     identity_private_key_hex = os.environ.get(
         HELPER_IDENTITY_PRIVATE_KEY_ENVIRONMENT_VARIABLE,
@@ -1875,9 +816,13 @@ async def run() -> None:
         disable_validator_service,
     )
     connection_manager = await QuicConnectionManager.create(identity_key)
-    metadata["bootnode_enr"] = build_helper_bootnode_enr(identity_key)
+    metadata["bootnode_enr"] = build_helper_bootnode_enr(
+        identity_key,
+        helper_p2p_port(),
+    )
     metadata["bootnode_qlean_enr"] = build_helper_bootnode_enr(
         identity_key,
+        helper_p2p_port(),
         include_udp=False,
     )
     metadata["bootnode_multiaddr"] = build_helper_bootnode_multiaddr(
@@ -1886,7 +831,13 @@ async def run() -> None:
     event_source = await LiveNetworkEventSource.create(connection_manager=connection_manager)
     event_source._hive_identify_response = build_identify_response(identity_key)
     configure_event_source_network(event_source, helper_gossip_fork_digest())
-    subscribe_gossip_topics(event_source, validator_registry, is_aggregator)
+    subscribe_gossip_topics(
+        event_source,
+        validator_registry,
+        is_aggregator,
+        helper_gossip_fork_digest(),
+        include_committee_aggregation=True,
+    )
     logger.info("Helper peer_id=%s", connection_manager.peer_id)
 
     node = Node.from_genesis(
@@ -1951,10 +902,10 @@ async def run() -> None:
         "config": ApiServerConfig(port=helper_api_port()),
         "store_getter": lambda: node.sync_service.store,
     }
-    if api_server_supports_signed_block_getter():
+    if api_server_supports_signed_block_getter(ApiServer):
         api_server_kwargs["signed_block_getter"] = lookup_published_block
     api_server = ApiServer(**api_server_kwargs)
-    metadata_runner = await start_metadata_server(metadata)
+    metadata_runner = await start_metadata_server(metadata, helper_metadata_port())
     refresh_status(event_source, node)
 
     if node.validator_service is not None:
@@ -1971,7 +922,7 @@ async def run() -> None:
     event_source.set_block_by_slot_lookup(lookup_reqresp_block_by_slot)
     event_source.set_current_slot_lookup(current_known_slot)
 
-    listener_task = await start_listener_and_gossipsub(event_source)
+    listener_task = await start_listener_and_gossipsub(event_source, listen_addr())
     await dial_bootnodes(event_source, bootnodes)
 
     node_task = asyncio.create_task(
@@ -2025,9 +976,10 @@ async def print_bootnode_metadata() -> None:
         json.dumps(
             {
                 "peer_id": peer_id,
-                "enr": build_helper_bootnode_enr(identity_key),
+                "enr": build_helper_bootnode_enr(identity_key, helper_p2p_port()),
                 "qlean_enr": build_helper_bootnode_enr(
                     identity_key,
+                    helper_p2p_port(),
                     include_udp=False,
                 ),
                 "multiaddr": build_helper_bootnode_multiaddr(peer_id),
